refactoring the todoList files
